Reorganize the `pg` package.  Spawn `pg/types.go` for everything not LSN.
